+++ mtui/commands/export.py
@@ -6,10 +6,10 @@ from pathlib import Path
 from ..cli.argparse import ArgumentParser
 from ..cli.completion import complete_choices_filelist
 from ..connector.qem_dashboard import DashboardAutoOpenQA
-from ..export import AutoExport, KernelExport, ManualExport
-from ..export.base import BaseExport
 from ..support.misc import requires_update
 from ..types.filelist import FileList
+from ..update_workflow.export import AutoExport, KernelExport, ManualExport
+from ..update_workflow.export.base import BaseExport
 from . import Command
 
 logger = getLogger("mtui.commands.export")
