+++ mtui/update_workflow/export/kernel.py
@@ -4,8 +4,8 @@ from datetime import datetime
 from logging import getLogger
 from pathlib import Path
 
-from ..support.fileops import ensure_dir_exists
-from ..types import FileList
+from ...support.fileops import ensure_dir_exists
+from ...types import FileList
 from .base import BaseExport
 from .downloader import download_logs
