+++ mtui/update_workflow/export/overview_inject.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 from collections.abc import MutableSequence
 from logging import getLogger
 
-from ..connector.oqa_search import (
+from ...connector.oqa_search import (
     OVERVIEW_BEGIN_MARKER,
     OVERVIEW_END_MARKER,
     BuildCheckResult,
